Fix build warning

+++ src/IRCClient.cpp
@@ -12,7 +12,7 @@
 std::vector<std::string> split(std::string const& text, char sep)
 {
     std::vector<std::string> tokens;
-    int start = 0, end = 0;
+    size_t start = 0, end = 0;
     while ((end = text.find(sep, start)) != std::string::npos)
     {
         tokens.push_back(text.substr(start, end - start));
